feat #23 - Move FloatingPanel Code & fpc dismiss Tap another annotation
